backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
whitenoise==6.6.0
